backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pydantic==2.7.0
scikit-learn==1.4.2
pandas==2.2.2
joblib==1.4.2
python-dotenv==1.0.1
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
twilio==9.2.3
apscheduler==3.10.4
python-dateutil==2.8.2
groq==0.5.0
